requirements.txt
python-telegram-bot>=20.0
minio>=7.2.0
python-dotenv>=1.0.0
requests>=2.31.0
